+++ public/echopay/EchopayFlash/LogoMC/LogoMC.tsx
@@ -1,24 +1,26 @@
 
 import React from 'react';
-import { useTheme } from '@mui/material';
+// import { useTheme } from '@mui/material';
 
 const Logo = () => {
-    const theme = useTheme();
-    const fill = theme.palette.primary.main;
+    // const theme = useTheme();
+    // const fill = theme.palette.primary.main;
     return (
-        <svg width="3163px" height="1064px" viewBox="0 0 3163 1064">
-            <g stroke="none" fill="none" fillRule="evenodd" strokeWidth="1">
-                <g transform="translate(81.0811, 81.0811)" fill="#FFFFFF">
-
-                    <g transform="translate(960.1087, 136.7732)" fillRule="nonzero">
-                        <path d="M12.8642962,321.607405 L12.8642962,70.7536291 L205.507132,70.7536291 L205.507132,133.145466 L77.1857773,133.145466 L77.1857773,169.165495 L167.557458,169.165495 L167.557458,227.376435 L77.1857773,227.376435 L77.1857773,259.537176 L205.507132,259.537176 L205.507132,321.607405 L12.8642962,321.607405 L12.8642962,321.607405 Z M394.387161,326.431516 C358.045524,326.431516 326.849606,313.56722 300.799406,288.160235 C274.427599,262.75325 261.241695,231.878939 261.241695,196.180517 C261.241695,160.482095 274.427599,129.607784 300.799406,104.200799 C326.849606,78.7938143 358.045524,65.9295181 394.387161,65.9295181 C431.050405,65.9295181 462.567931,78.7938143 488.618131,104.200799 C508.236182,123.175636 520.457264,145.366547 524.959768,170.130317 L457.74382,170.130317 C454.527746,161.768525 449.382027,154.049947 442.949879,147.617799 C429.763976,134.431895 413.683605,127.999747 394.387161,127.999747 C375.412324,127.999747 359.331954,134.753503 346.14605,147.939406 C332.960147,161.12531 326.206391,177.20568 326.206391,196.180517 C326.206391,215.155354 332.960147,231.235724 346.14605,244.421628 C359.331954,257.607532 375.412324,264.361287 394.387161,264.361287 C413.683605,264.361287 429.763976,257.607532 442.949879,244.421628 C451.954887,235.416621 457.74382,225.446791 460.638287,213.547317 L526.246197,213.547317 C522.708516,242.170376 510.487434,266.934146 488.618131,288.160235 C462.567931,313.56722 431.050405,326.431516 394.387161,326.431516 Z M581.98076,321.607405 L581.98076,70.7536291 L645.980634,70.7536291 L645.980634,168.200673 L748.251789,168.200673 L748.251789,70.7536291 L812.57327,70.7536291 L812.57327,321.607405 L748.251789,321.607405 L748.251789,228.341258 L645.980634,228.341258 L645.980634,321.607405 L581.98076,321.607405 Z M1001.4533,326.431516 C965.111662,326.431516 933.915744,313.56722 907.865544,288.160235 C881.493737,262.75325 868.307833,231.878939 868.307833,196.180517 C868.307833,160.482095 881.493737,129.607784 907.865544,104.200799 C933.915744,78.7938143 965.111662,65.9295181 1001.4533,65.9295181 C1038.11654,65.9295181 1069.63407,78.7938143 1095.68427,104.200799 C1121.73447,129.607784 1134.59876,160.482095 1134.59876,196.180517 C1134.59876,231.878939 1121.73447,262.75325 1095.68427,288.160235 C1069.63407,313.56722 1038.11654,326.431516 1001.4533,326.431516 Z M1001.4533,264.361287 C1020.74974,264.361287 1036.83011,257.607532 1050.01602,244.421628 C1063.20192,231.235724 1069.63407,215.155354 1069.63407,196.180517 C1069.63407,177.20568 1063.20192,160.803703 1050.01602,147.617799 C1036.83011,134.431895 1020.74974,127.999747 1001.4533,127.999747 C982.478462,127.999747 966.398092,134.753503 953.212188,147.939406 C940.026285,161.12531 933.272529,177.20568 933.272529,196.180517 C933.272529,215.155354 940.026285,231.235724 953.212188,244.421628 C966.398092,257.607532 982.478462,264.361287 1001.4533,264.361287 L1001.4533,264.361287 Z M1190.33333,321.607405 L1190.33333,70.7536291 L1300.96628,70.7536291 C1332.4838,70.7536291 1356.92596,78.7938143 1373.97116,94.8741845 C1391.01635,111.276162 1399.37814,131.537429 1399.37814,155.014769 C1399.37814,179.77854 1391.01635,200.361413 1373.97116,216.441784 C1356.92596,232.522154 1332.80541,240.883947 1300.96628,240.883947 L1254.65481,240.883947 L1254.65481,321.607405 L1190.33333,321.607405 Z M1254.65481,189.426762 L1297.42859,189.426762 C1322.19236,189.426762 1334.41345,178.170502 1334.41345,155.336377 C1334.41345,145.688155 1331.51898,137.969577 1325.40844,131.537429 C1319.2979,125.105281 1309.97128,122.210814 1297.42859,122.210814 L1254.65481,122.210814 L1254.65481,189.426762 Z M1455.1127,321.607405 L1556.74064,70.7536291 L1620.41891,70.7536291 L1721.72524,321.607405 L1651.61483,321.607405 L1631.35356,265.647717 L1546.1276,265.647717 L1525.22312,321.607405 L1455.1127,321.607405 Z M1566.06726,213.547317 L1610.77069,213.22571 L1588.57978,153.085125 L1566.06726,213.547317 Z M1876.51489,321.607405 L1876.51489,229.627687 L1777.45981,70.7536291 L1853.03755,70.7536291 L1908.35402,161.768525 L1963.9921,70.7536291 L2039.89145,70.7536291 L1940.51476,229.627687 L1940.51476,321.607405 L1876.51489,321.607405 Z" id="Shape"></path>
-                    </g>
-                    <g transform="translate(971.7121, 598.2334)" fillRule="nonzero">
-                        <path d="M58.0013983,27.7397992 L58.0013983,40.5427835 L37.2450451,40.5427835 L37.2450451,96.9923049 L22.0172532,96.9923049 L22.0172532,40.5427835 L1.26089996,40.5427835 L1.26089996,27.7397992 L58.0013983,27.7397992 Z M105.374865,27.7397992 L105.374865,54.3156908 L133.405641,54.3156908 L133.405641,27.7397992 L148.633433,27.7397992 L148.633433,96.9923049 L133.405641,96.9923049 L133.405641,67.118675 L105.374865,67.118675 L105.374865,96.9923049 L90.147073,96.9923049 L90.147073,27.7397992 L105.374865,27.7397992 Z M238.004567,27.7397992 L238.004567,40.5427835 L201.438468,40.5427835 L201.438468,55.3826061 L234.997806,55.3826061 L234.997806,67.2156673 L201.438468,67.2156673 L201.438468,84.1893207 L238.780506,84.1893207 L238.780506,96.9923049 L186.210677,96.9923049 L186.210677,27.7397992 L238.004567,27.7397992 Z M333.236462,27.7397992 L344.196592,74.8780594 L344.390577,74.8780594 L356.417623,27.7397992 L370.675492,27.7397992 L382.508553,75.4600132 L382.702538,75.4600132 L394.050637,27.7397992 L408.987452,27.7397992 L390.36493,96.9923049 L375.23413,96.9923049 L363.498061,49.8540447 L363.304077,49.8540447 L351.761992,96.9923049 L336.340216,96.9923049 L318.00867,27.7397992 L333.236462,27.7397992 Z M472.946603,27.7397992 L498.843548,96.9923049 L483.033803,96.9923049 L477.796218,81.5705284 L451.899273,81.5705284 L446.467704,96.9923049 L431.045927,96.9923049 L457.330842,27.7397992 L472.946603,27.7397992 L472.946603,27.7397992 Z M464.896242,44.8104449 L455.875957,70.2224288 L473.819534,70.2224288 L465.090226,44.8104449 L464.896242,44.8104449 Z M531.474184,27.7397992 L547.671899,55.0916292 L563.772622,27.7397992 L580.746275,27.7397992 L555.043315,70.4164134 L555.043315,96.9923049 L539.815523,96.9923049 L539.815523,70.0284442 L514.306546,27.7397992 L531.474184,27.7397992 Z M713.223215,27.7397992 L713.223215,40.5427835 L692.466862,40.5427835 L692.466862,96.9923049 L677.23907,96.9923049 L677.23907,40.5427835 L656.482717,40.5427835 L656.482717,27.7397992 L713.223215,27.7397992 Z M776.406427,26.09093 C786.622627,26.09093 794.851454,29.4856607 801.090969,36.275122 C807.330484,43.0645834 810.450726,51.8588757 810.450726,62.657029 C810.450726,73.1971828 807.314674,81.8294979 801.042473,88.5539744 C794.770271,95.2784509 786.558612,98.6411741 776.406427,98.6411741 C766.190228,98.6411741 757.961401,95.2787419 751.721886,88.5539744 C745.482371,81.8292069 742.362128,73.1971828 742.362128,62.657029 C742.362128,51.9228906 745.49818,43.145087 751.770382,36.3236182 C758.042583,29.5021494 766.254243,26.09093 776.406427,26.09093 L776.406427,26.09093 Z M757.58992,62.657029 C757.58992,69.6404749 759.238692,75.2495399 762.536528,79.4851939 C765.834363,83.7208478 770.457404,85.8381899 776.406427,85.8381899 C782.225966,85.8381899 786.816611,83.7528553 790.179335,79.5821862 C793.542058,75.4115171 795.222934,69.7694747 795.222934,62.657029 C795.222934,55.2211139 793.557868,49.4010907 790.227831,45.1984141 C786.897794,40.9957375 782.258458,38.8939143 776.309435,38.8939143 C770.489897,38.8939143 765.91477,40.9954465 762.585024,45.1984141 C759.255278,49.4013816 757.58992,55.220629 757.58992,62.657029 Z M927.505889,27.7397992 C934.877305,27.7397992 940.696843,29.7443392 944.964504,33.7533221 C949.232166,37.7623051 951.365996,43.1615757 951.365996,49.951037 C951.365996,56.8049983 949.232166,62.2205636 944.964504,66.1972481 C940.696843,70.1739326 934.909312,72.1622749 927.602882,72.1622749 L911.405167,72.1622749 L911.405167,96.9923049 L896.177375,96.9923049 L896.177375,27.7397992 L927.505889,27.7397992 Z M923.529205,60.3292137 C927.926157,60.3292137 931.207601,59.5209768 933.373924,57.904406 C935.540247,56.2878353 936.623166,53.6367446 936.623166,49.951037 C936.623166,46.3946202 935.572449,43.7760219 933.470916,42.0946603 C931.369384,40.4132987 928.087843,39.5728604 923.626197,39.5728604 L911.405167,39.5728604 L911.405167,60.3292137 L923.529205,60.3292137 L923.529205,60.3292137 Z M1012.41538,27.7397992 L1038.31232,96.9923049 L1022.50258,96.9923049 L1017.26499,81.5705284 L991.368048,81.5705284 L985.936479,96.9923049 L970.514702,96.9923049 L996.799617,27.7397992 L1012.41538,27.7397992 L1012.41538,27.7397992 Z M1004.36502,44.8104449 L995.344732,70.2224288 L1013.28831,70.2224288 L1004.559,44.8104449 L1004.36502,44.8104449 L1004.36502,44.8104449 Z M1070.94296,27.7397992 L1087.14067,55.0916292 L1103.2414,27.7397992 L1120.21505,27.7397992 L1094.51209,70.4164134 L1094.51209,96.9923049 L1079.2843,96.9923049 L1079.2843,70.0284442 L1053.77532,27.7397992 L1070.94296,27.7397992 Z M1252.69199,27.7397992 L1252.69199,40.5427835 L1231.93564,40.5427835 L1231.93564,96.9923049 L1216.70785,96.9923049 L1216.70785,40.5427835 L1195.95149,40.5427835 L1195.95149,27.7397992 L1252.69199,27.7397992 Z M1315.8752,26.09093 C1326.0914,26.09093 1334.32023,29.4856607 1340.55974,36.275122 C1346.79926,43.0645834 1349.9195,51.8588757 1349.9195,62.657029 C1349.9195,73.1971828 1346.78345,81.8294979 1340.51125,88.5539744 C1334.23905,95.2784509 1326.02739,98.6411741 1315.8752,98.6411741 C1305.659,98.6411741 1297.43018,95.2787419 1291.19066,88.5539744 C1284.95115,81.8292069 1281.8309,73.1971828 1281.8309,62.657029 C1281.8309,51.9228906 1284.96696,43.145087 1291.23916,36.3236182 C1297.51136,29.5021494 1305.72302,26.09093 1315.8752,26.09093 L1315.8752,26.09093 Z M1297.0587,62.657029 C1297.0587,69.6404749 1298.70747,75.2495399 1302.0053,79.4851939 C1305.30314,83.7208478 1309.92618,85.8381899 1315.8752,85.8381899 C1321.69474,85.8381899 1326.28539,83.7528553 1329.64811,79.5821862 C1333.01083,75.4115171 1334.69171,69.7694747 1334.69171,62.657029 C1334.69171,55.2211139 1333.02664,49.4010907 1329.69661,45.1984141 C1326.36657,40.9957375 1321.72723,38.8939143 1315.77821,38.8939143 C1309.95867,38.8939143 1305.38354,40.9954465 1302.0538,45.1984141 C1298.72405,49.4013816 1297.0587,55.220629 1297.0587,62.657029 L1297.0587,62.657029 Z M1414.16963,27.7397992 C1423.99786,27.7397992 1431.72524,30.7788591 1437.35079,36.8570759 C1442.97634,42.9352926 1445.78912,51.3089293 1445.78912,61.9780828 C1445.78912,73.0352056 1442.96015,81.6355133 1437.30229,87.778036 C1431.64444,93.9205586 1424.06284,96.9923049 1414.5576,96.9923049 L1384.48998,96.9923049 L1384.48998,27.7397992 L1414.16963,27.7397992 Z M1413.19971,84.1893207 C1418.69577,84.1893207 1422.96392,82.3787653 1426.00269,78.7577516 C1429.04146,75.1367379 1430.56133,69.9959517 1430.56133,63.3359751 C1430.56133,55.7060754 1428.92866,49.9995332 1425.66322,46.2168333 C1422.39778,42.4341334 1417.37038,40.5427835 1410.58091,40.5427835 L1399.71778,40.5427835 L1399.71778,84.1893207 L1413.19971,84.1893207 L1413.19971,84.1893207 Z M1515.08285,27.7397992 L1540.97979,96.9923049 L1525.17005,96.9923049 L1519.93246,81.5705284 L1494.03552,81.5705284 L1488.60395,96.9923049 L1473.18217,96.9923049 L1499.46709,27.7397992 L1515.08285,27.7397992 L1515.08285,27.7397992 Z M1507.03249,44.8104449 L1498.0122,70.2224288 L1515.95578,70.2224288 L1507.22647,44.8104449 L1507.03249,44.8104449 L1507.03249,44.8104449 Z M1573.61043,27.7397992 L1589.80814,55.0916292 L1605.90887,27.7397992 L1622.88252,27.7397992 L1597.17956,70.4164134 L1597.17956,96.9923049 L1581.95177,96.9923049 L1581.95177,70.0284442 L1556.44279,27.7397992 L1573.61043,27.7397992 Z" ></path>
-                    </g>
-                </g>
-            </g>
-        </svg>
+        <>
+            <img
+                src={'/echopay/svg/logo.svg'}
+                alt="Logo"
+                style={{
+                    display: 'block',
+                    maxWidth: '100%',
+                    maxHeight: '100%',
+                    width: 'auto',
+                    height: 'auto',
+                    margin: '0 auto',
+                    objectFit: 'contain',
+                }}
+            />
+        </>
     );
 };
